+++ src/hash/formats.rs
@@ -2,5 +2,3 @@ pub mod json;
 pub mod rdf;
 pub mod text;
 pub mod xml;
-
-mod providers;
