+++ src/hash/formats/xml/provisions/libxml.rs
@@ -1,0 +1,9 @@
+use super::{Error, Result};
+
+pub struct LibXML;
+
+impl super::XMLProvider for LibXML {
+    fn find_xpath(source: &str, xpath: &str) -> Result<String> {
+        unimplemented!()
+    }
+}
