+++ src/hash/formats/xml/provisions/other.rs
@@ -1,0 +1,11 @@
+mod sxd;
+
+use super::Result;
+
+pub struct OtherProviders;
+
+impl super::XMLProvider for OtherProviders {
+    fn find_xpath(source: &str, xpath: &str) -> Result<String> {
+        sxd::find_xpath(source, xpath)
+    }
+}
