+++ src/hash/formats/xml/provisions/other/sxd.rs
@@ -1,4 +1,9 @@
-pub fn find_xpath(source: &str, xpath: &str) -> Option<String> {
+use sxd_document::parser;
+use sxd_xpath::{evaluate_xpath, Value};
+
+use super::super::{Error, Result};
+
+pub fn find_xpath(source: &str, xpath: &str) -> Result<String> {
     let pkg = parser::parse(source.as_ref())?;
     let doc = pkg.as_document();
 
@@ -6,12 +11,12 @@ pub fn find_xpath(source: &str, xpath: &str) -> Option<String> {
 
     let hash = match value {
         Value::Nodeset(nodes) => {
-            let node = nodes.iter().last().ok_or(XMLError::NotFound)?;
+            let node = nodes.iter().last().ok_or(Error::NotFound)?;
 
             node.string_value()
         }
         Value::String(text) => text,
-        _ => return Err(XMLError::InvalidValue),
+        _ => return Err(Error::InvalidValue),
     };
 
     Ok(hash)
